+++ rust/operator-binary/src/discovery.rs
@@ -1,18 +1,19 @@
 //! Discovery for Druid.  We make Druid discoverable by putting a connection string to the router service
 //! inside a config map.  We only provide a connection string to the router service, since it serves as
 //! a gateway to the cluster for client queries.
-use snafu::{OptionExt, ResultExt, Snafu};
+use snafu::{ResultExt, Snafu};
 use stackable_operator::{
     builder::{configmap::ConfigMapBuilder, meta::ObjectMetaBuilder},
     commons::product_image_selection::ResolvedProductImage,
+    crd::listener::v1alpha1::Listener,
     k8s_openapi::api::core::v1::ConfigMap,
     kube::{Resource, ResourceExt, runtime::reflector::ObjectRef},
-    utils::cluster_info::KubernetesClusterInfo,
 };
 
 use crate::{
     DRUID_CONTROLLER_NAME,
     crd::{DruidRole, build_recommended_labels, security::DruidTlsSecurity, v1alpha1},
+    listener::build_listener_connection_string,
 };
 
 #[derive(Snafu, Debug)]
@@ -35,24 +36,27 @@ pub enum Error {
     AddRecommendedLabels {
         source: stackable_operator::builder::meta::Error,
     },
+
+    #[snafu(display("failed to configure listener discovery configmap"))]
+    ListenerConfiguration { source: crate::listener::Error },
 }
 
 /// Builds discovery [`ConfigMap`]s for connecting to a [`v1alpha1::DruidCluster`].
 pub async fn build_discovery_configmaps(
     druid: &v1alpha1::DruidCluster,
     owner: &impl Resource<DynamicType = ()>,
-    cluster_info: &KubernetesClusterInfo,
     resolved_product_image: &ResolvedProductImage,
     druid_tls_security: &DruidTlsSecurity,
+    listener: Listener,
 ) -> Result<Vec<ConfigMap>, Error> {
     let name = owner.name_unchecked();
     Ok(vec![build_discovery_configmap(
         druid,
         owner,
-        cluster_info,
         resolved_product_image,
         druid_tls_security,
         &name,
+        listener,
     )?])
 }
 
@@ -60,22 +64,17 @@ pub async fn build_discovery_configmaps(
 fn build_discovery_configmap(
     druid: &v1alpha1::DruidCluster,
     owner: &impl Resource<DynamicType = ()>,
-    cluster_info: &KubernetesClusterInfo,
     resolved_product_image: &ResolvedProductImage,
     druid_tls_security: &DruidTlsSecurity,
     name: &str,
+    listener: Listener,
 ) -> Result<ConfigMap, Error> {
-    let router_host = format!(
-        "{}:{}",
-        druid
-            .role_service_fqdn(&DruidRole::Router, cluster_info)
-            .with_context(|| NoServiceFqdnSnafu)?,
-        if druid_tls_security.tls_enabled() {
-            DruidRole::Router.get_https_port()
-        } else {
-            DruidRole::Router.get_http_port()
-        }
-    );
+    let router_host = build_listener_connection_string(
+        listener,
+        druid_tls_security,
+        &DruidRole::Router.to_string(),
+    )
+    .context(ListenerConfigurationSnafu)?;
     let sqlalchemy_conn_str = format!("druid://{}/druid/v2/sql", router_host);
     let avatica_conn_str = format!(
         "jdbc:avatica:remote:url=http://{}/druid/v2/sql/avatica/",