+++ rust/operator-binary/src/main.rs
@@ -34,8 +34,10 @@ mod discovery;
 mod druid_controller;
 mod extensions;
 mod internal_secret;
+mod listener;
 mod operations;
 mod product_logging;
+mod service;
 
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));